requirements.txt
Pillow
numpy
numba
opencv-python
scipy
scikit-image
scikit-learn
tqdm
tensorboard
beautifultable
lpips
onnx_runtime
tensorflow-gpu
git+https://github.com/onnx/onnx-tensorflow.git